Feel-leaf-deploy/requirements.txt
flask
flask-cors
firebase-admin
google-generativeai
